backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
gunicorn==21.2.0
django-redis==5.4.0
redis==5.0.1